requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
arxiv==2.1.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
feedparser==6.0.11
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.23
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
llama-cloud==0.1.6
llama-index==0.12.5
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.5
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.4
llama-index-multi-modal-llms-openai==0.4.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.1
llama-parse==0.5.17
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.16.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.57.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pinecone==5.4.1
pinecone-plugin-assistant==0.4.1
pinecone-plugin-interface==0.0.7
propcache==0.2.1
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.1.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.40.2
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.2.3
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.3
